convert funcs

+++ uint32.go
@@ -9,6 +9,6 @@ func Uint32ToBytes(i uint32) []byte {
 	}
 }
 
-func Uint32FromBytes(b []byte) int32 {
-	return int32(uint32(b[3]) | uint32(b[2])<<8 | uint32(b[1])<<16 | uint32(b[0])<<24)
+func Uint32FromBytes(b []byte) uint32 {
+	return uint32(b[3]) | uint32(b[2])<<8 | uint32(b[1])<<16 | uint32(b[0])<<24
 }
